refactored for versioning

+++ create-scripts/delete-cluster-and-poc.sh
@@ -5,9 +5,17 @@ app_namespace=event-poc
 cluster_name=event-driven-poc
 #cluster_name=$(eksctl get cluster -o json | jq -r ".[0].Name")
 
-kubectl delete -f manifests/
+VERS=$1
 
-kubectl delete -f rabbitmq/rabbitmq-cluster.yaml
+if [ $# -eq 0 ] || [ $# -gt 1 ]; then
+  echo "Zero or too many arguments supplied..."
+  echo "Example: ./create-scripts/delete-cluster-and-poc.sh v1"
+  exit 1
+fi
+
+kubectl delete -f manifests/${VERS}
+
+kubectl delete -f rabbitmq/${VERS}/rabbitmq-cluster.yaml
 
 aws cloudformation delete-stack --stack-name $cluster_name-efs
